add types base file
