--- lib/file-writer.ts
@@ -1,15 +1,0 @@
-import fs from 'fs';
-import logger from './logger';
-
-export function writeTypesToFile(types: string[], outputPath: string): void {
-  // Ensure the output directory exists
-  const outputDir = outputPath.split('/').slice(0, -1).join('/');
-  fs.mkdirSync(outputDir, { recursive: true });
-
-  let toWrite = `// This file was generated by the openapi-to-typescript tool.\n`;
-  toWrite += `// Do not modify it by hand.\n\n`;
-  toWrite += types.join('\n');
-  toWrite += '\n';
-  fs.writeFileSync(outputPath, toWrite);
-  logger.info(`wrote types to ${outputPath}`);
-}
