+++ lib/writer/file-writer.ts
@@ -1,0 +1,35 @@
+import fs from 'fs';
+import logger from '../logger';
+
+export function writeTypesToFile(types: string[], outputPath: string): void {
+  // Make output directory if it doesn't exist
+  const outputDir = outputPath.split('/').slice(0, -1).join('/');
+  fs.mkdirSync(outputDir, { recursive: true });
+
+  // Load the types base file, removing initial comments
+  const typesBasePath = `${__dirname}/types-base.ts`;
+  const typesBase = fs.readFileSync(typesBasePath, 'utf8').split('\n');
+
+  let toWrite = `// This file was generated by the openapi-to-typescript tool.\n`;
+  toWrite += `// Do not modify it by hand.\n\n`;
+
+  // Find the index where the comments end and types base starts.
+  let i = 0;
+  for (; i < typesBase.length; i++) {
+    const lineContents = typesBase[i].trim();
+    if (!lineContents.startsWith('//') && lineContents !== '') {
+      break;
+    }
+  }
+  const typesBaseContents = typesBase.slice(i).join('\n');
+
+  toWrite += '// START types-base.ts\n';
+  toWrite += typesBaseContents;
+  toWrite += '// END types-base.ts\n\n';
+
+  // Append the types to the output file
+  toWrite += types.join('\n');
+  toWrite += '\n';
+  fs.writeFileSync(outputPath, toWrite);
+  logger.info(`wrote types to ${outputPath}`);
+}
