+++ lib/writer/types-base.ts
@@ -1,0 +1,4 @@
+// This is the types base file, which contains some common code used throughout
+// the project. All types will be added below the contents of this file.
+
+type Nullable<T> = T | null;
